+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased Changes]
 
+- add .lazy_attr_accessor
+
 ## 0.1.0
 
 - initial release