+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased Changes]
 
+## 0.2.0
+
 - add .lazy_attr_accessor
 - .lazy_attr_reader and .lazy_attr_accessor return the same value as their normal counterparts
 - #eager_load_attributes! returns the names of only the  attributes which were eager-loaded